+++ test/sample-sync-doc-definitions-spec.js
@@ -504,6 +504,7 @@ describe('The sample-sync-doc-definitions sync function', function() {
         subject: '', // missing sender, empty subject
         'whatsthis?': 'something I dont recognize!', // unrecognized property
         createdAt: '2016-02-29T25:13:43.666Z', // invalid hour
+        firstReadAt: '201-07-14T21:21:21.212-08:00', // invalid year
         actions: [ { url: 24 }, null ] // integer url, non-existent label
       };
 
@@ -519,7 +520,8 @@ describe('The sample-sync-doc-definitions sync function', function() {
           'item "actions[0].url" must be a string',
           'required item "actions[0].label" is missing',
           'required item "actions[1]" is missing',
-          'property "whatsthis?" is not supported'
+          'property "whatsthis?" is not supported',
+          'item "firstReadAt" must be an ISO 8601 date string with optional time and time zone components'
         ]);
     });
 
@@ -531,6 +533,7 @@ describe('The sample-sync-doc-definitions sync function', function() {
         subject: 'a different subject',
         message: 'last warning!',
         createdAt: '2016-02-29T17:13:43.666Z',
+        firstReadAt: '2016-07-14T21:21:21.212-08:00',
         actions: [ { url: 'http://foobar.baz/lastwarning', label: 'pay up here'} ]
       };
       var oldDoc = {
@@ -552,6 +555,7 @@ describe('The sample-sync-doc-definitions sync function', function() {
         sender: '', // missing type, empty sender
         message: '', // missing subject, empty message
         createdAt: '2016-04-29T17:13:43.666Z', // changed createdAt
+        firstReadAt: '2016-07-14T21:24:16.997-08:00',
         actions: [ { label: ''} ]
       };
       var oldDoc = { // valid oldDoc
@@ -561,6 +565,7 @@ describe('The sample-sync-doc-definitions sync function', function() {
         subject: 'a different subject',
         message: 'last warning!',
         createdAt: '2016-02-29T17:13:43.666Z',
+        firstReadAt: '2016-07-14T21:21:21.212-08:00',
         actions: [ { url: 'http://foobar.baz/lastwarning', label: 'pay up here'} ]
       };
 
@@ -580,7 +585,8 @@ describe('The sample-sync-doc-definitions sync function', function() {
           'value of item "createdAt" may not be modified',
           'value of item "actions" may not be modified',
           'required item "actions[0].url" is missing',
-          'item "actions[0].label" must not be empty'
+          'item "actions[0].label" must not be empty',
+          'value of item "firstReadAt" may not be modified'
         ]);
     });
 
@@ -944,6 +950,7 @@ describe('The sample-sync-doc-definitions sync function', function() {
         _id: 'biz.275.notification.ABC.processedTransport.XYZ',
         processedBy: 'foobar',
         processedAt: '2016-06-03T21:02:19.013Z',
+        sentAt: '2016-07-15'
       };
 
       verifyProcessingSummaryNotWritten(
@@ -954,7 +961,8 @@ describe('The sample-sync-doc-definitions sync function', function() {
           'item "nonce" must be a string',
           'value of item "processedBy" may not be modified',
           'value of item "processedAt" may not be modified',
-          'item "sentAt" must be an ISO 8601 date string with optional time and time zone components'
+          'item "sentAt" must be an ISO 8601 date string with optional time and time zone components',
+          'value of item "sentAt" may not be modified'
         ]);
     });